refactor(svgElements): extracted all basic svg element parser to a seperate module
